shapes

+++ PontCoopScheduler.h
@@ -17,8 +17,8 @@
 
 
 typedef struct {
-  uint32_t delay;
-  uint32_t period;
+  uint16_t delay;
+  uint16_t period;
   uint8_t run;
   void (*exec)(void *handle);
   void *handle;
